+++ kernel-lib/src/task/mod.rs
@@ -4,10 +4,9 @@ pub fn terminate(code: i32) -> ! {
     unsafe {
         core::arch::asm! {
             "int 80h",
-            in("eax") syscall::Request::TerminateCurrentTask,
+            in("eax") syscall::Request::TerminateCurrentTask as u32,
             in("edx") code,
             options(noreturn)
         }
     }
 }
-
